Fix PickItem, SetItem and Remove

+++ src/Neo.VM/JumpTable/JumpTable.Compound.cs
@@ -372,10 +372,10 @@ partial class JumpTable
         {
             case VMArray array:
                 {
-                    var index = (int)key.GetInteger();
+                    var index = key.GetInteger();
                     if (index < 0 || index >= array.Count)
                         throw new CatchableException($"The index of {nameof(VMArray)} is out of range, {index}/[0, {array.Count}).");
-                    engine.Push(array[index]);
+                    engine.Push(array[(int)index]);
                     break;
                 }
             case Map map:
@@ -388,18 +388,18 @@ partial class JumpTable
             case PrimitiveType primitive:
                 {
                     var byteArray = primitive.GetSpan();
-                    var index = (int)key.GetInteger();
+                    var index = key.GetInteger();
                     if (index < 0 || index >= byteArray.Length)
                         throw new CatchableException($"The index of {nameof(PrimitiveType)} is out of range, {index}/[0, {byteArray.Length}).");
-                    engine.Push((BigInteger)byteArray[index]);
+                    engine.Push((BigInteger)byteArray[(int)index]);
                     break;
                 }
             case Buffer buffer:
                 {
-                    var index = (int)key.GetInteger();
+                    var index = key.GetInteger();
                     if (index < 0 || index >= buffer.Size)
                         throw new CatchableException($"The index of {nameof(Buffer)} is out of range, {index}/[0, {buffer.Size}).");
-                    engine.Push((BigInteger)buffer.InnerBuffer.Span[index]);
+                    engine.Push((BigInteger)buffer.InnerBuffer.Span[(int)index]);
                     break;
                 }
             default:
@@ -441,10 +441,10 @@ partial class JumpTable
         {
             case VMArray array:
                 {
-                    var index = (int)key.GetInteger();
+                    var index = key.GetInteger();
                     if (index < 0 || index >= array.Count)
                         throw new CatchableException($"The index of {nameof(VMArray)} is out of range, {index}/[0, {array.Count}).");
-                    array[index] = value;
+                    array[(int)index] = value;
                     break;
                 }
             case Map map:
@@ -454,15 +454,15 @@ partial class JumpTable
                 }
             case Buffer buffer:
                 {
-                    var index = (int)key.GetInteger();
+                    var index = key.GetInteger();
                     if (index < 0 || index >= buffer.Size)
                         throw new CatchableException($"The index of {nameof(Buffer)} is out of range, {index}/[0, {buffer.Size}).");
                     if (value is not PrimitiveType p)
                         throw new InvalidOperationException($"Only primitive type values can be set in {nameof(Buffer)} in {instruction.OpCode}.");
-                    var b = (int)p.GetInteger();
+                    var b = p.GetInteger();
                     if (b < sbyte.MinValue || b > byte.MaxValue)
                         throw new InvalidOperationException($"Overflow in {instruction.OpCode}, {b} is not a byte type.");
-                    buffer.InnerBuffer.Span[index] = (byte)b;
+                    buffer.InnerBuffer.Span[(int)index] = (byte)b;
                     buffer.InvalidateHashCode();
                     break;
                 }
@@ -511,10 +511,10 @@ partial class JumpTable
         switch (x)
         {
             case VMArray array:
-                var index = (int)key.GetInteger();
+                var index = key.GetInteger();
                 if (index < 0 || index >= array.Count)
                     throw new InvalidOperationException($"The index of {nameof(VMArray)} is out of range, {index}/[0, {array.Count}).");
-                array.RemoveAt(index);
+                array.RemoveAt((int)index);
                 break;
             case Map map:
                 map.Remove(key);
